+++ src/constants/tests.ts
@@ -26,9 +26,13 @@ export const TestIds = {
   groupsEditor: {
     buttonAddNew: 'data-testid fields-editor button-add-new',
     buttonRemove: 'data-testid fields-editor button-remove',
+    buttonStartRename: 'data-testid fields-editor button-start-rename',
+    buttonCancelRename: 'data-testid fields-editor button-cancel-rename',
+    buttonSaveRename: 'data-testid fields-editor button-save-rename',
+    fieldName: 'data-testid fields-editor field-name',
     item: (name: string) => `data-testid fields-editor item-${name}`,
     newItem: 'data-testid fields-editor new-level',
-    newItemName: 'fields-editor new-item-name',
+    newItemName: 'data-testid fields-editor new-item-name',
   },
   levelsEditor: {
     buttonAddNew: 'data-testid levels-editor button-add-new',